requirements.txt
certifi==2020.12.5
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.1
six==1.15.0
wincertstore==0.2
